Execute on tty0 virtual console
The old redirection seems to have been put in place because
plymouthd's shutdown code leaves the active console (tty63) in
unusable state.

To avoid this we can just block plymouth's shutdown screen from
running. Now we arrive on a usable console with no redirection
needed.

And in this cleaner state, we can work around an issue with early
EOS-2.5 EC-100 builds which had the serial console as the default
console: using tty0 will force the output to the screen, also
without breaking the default case for other images where tty0 is
already the default.

Fixes an issue where the text would not be visible on screen when
reflashing EC-100.