Use pipefail to handle decompressor failure
Switch to bash, and use the pipefail feature to detect when
a failure is returned from the decompressor or from the exit
condition in unz().

We can only do this in the first call to unz. The second one
will return an expected failure from gzip (and probably xz too) due
to broken pipe.