Stop processes before killing them
The previous approach to looping through /proc and killing the processes
with "kill -9" is racy in that one (not yet killed) process may spawn
another one while the loop is running, and hence will not be killed.
This was seen as the systemd-bootchart process still running and blocking
the flashing process.

To avoid such races we can first pause all processes before killing them.
Also kill's -1 argument is simpler here and avoids races.

Also remove a noisy debug 'ls' command which was obscuring the original
failure.