start.sh: Try to load the in-kernel exfat driver if needed
Prior to EOS 4.0 exFAT support was provided via FUSE. Since the initrd
does not contain the FUSE driver or the kernel module, it will not be
possible to mount the USB device containing the image once we pivot to
the initrd.

One way around this problem is to manually load the in-kernel exfat
driver before pivot-root, so it is available in the running kernel after
pivot-root (which is also how this works for any other filesystem we
support in the kernel).

This commit tried to load the in-kernel exfat driver if it detects any
exFAT filesystem attached to the system. This is the best we can do from
start.sh, as at the point it is running it has already been copied from
the USB device to /var by eos-factory-test-helper.service.

https://phabricator.endlessm.com/T31536